requirements_postgres.txt
requests
httpx
networkx
numpy
matplotlib
scikit-learn
fastapi
uvicorn
rapidfuzz
async-lru
sqlalchemy
psycopg2-binary
tqdm
